+++ base/src/main/java/io/spine/code/proto/FileSet.java
@@ -45,8 +45,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.collect.Maps.newHashMap;
 import static com.google.common.collect.Maps.newHashMapWithExpectedSize;
 import static com.google.common.flogger.LazyArgs.lazy;
-import static io.spine.code.FsObject.checkExists;
 import static io.spine.code.proto.Linker.link;
+import static io.spine.io.Files2.checkExists;
 import static io.spine.util.Exceptions.newIllegalStateException;
 import static java.lang.System.lineSeparator;
 import static java.util.stream.Collectors.joining;
